requirements.txt
# Shared dependencies for the SimpleStream codebase.
# Install torch / torchvision separately for your platform first
# (see https://pytorch.org/get-started/locally/), then install one of:
#   - requirements-qwen3.txt   (Qwen3-VL stack)
#   - requirements-qwen25.txt  (Qwen2.5-VL stack)

# Video decoding
av>=14.4.0

# Image / array utilities
Pillow>=10.0.0
numpy>=1.24.0

# Progress bars
tqdm>=4.65.0

# Qwen-VL preprocessing helper (used by both backbones)
qwen-vl-utils==0.0.14
